refactor(semantic): extract embedder runtime helpers into runtime.rs
Context: A new RAG layer (reranker + LLM + chat UI) is being added on
top of the existing semantic search system. The RAG code needs to reuse
the embedder lazy-load/idle-unload machinery and the deserialized chunk
cache that currently live as module-private items inside
`commands/semantic.rs`. Without a refactor, the new modules would either
duplicate the logic or be forced to live inside the same file.

Solution: Moved the five helpers (`invalidate_search_cache`,
`get_or_load_cache`, `unload_embedder`, `ensure_embedder_loaded`,
`schedule_embedder_unload`), the `CachedChunk` struct, and the static
state they manage (`EMBEDDER`, `VAULT_PATH`, `UNLOAD_GENERATION`,
`EMBEDDER_IDLE_TIMEOUT_SECS`, `SEARCH_CACHE`) into a new
`semantic/runtime.rs` module with `pub(crate)` visibility.
`commands/semantic.rs` imports them via
`use crate::semantic::runtime::{...}`. `BUILD_LOCK` stays in
`commands/semantic.rs` because only `build_semantic_index` uses it.

Behavior: No user-visible or behavioral change. All commands
(`init_semantic_search`, `build_semantic_index`, `search_semantic`,
`update_semantic_file`, `shutdown_semantic`, `get_semantic_stats`,
`get_semantic_file_status`) continue to use the same static state and
helpers through the new import path. Future RAG modules
(`rag/retrieval.rs`, `commands/rag.rs`) will reuse the same helpers
without duplicating logic or violating module boundaries.

Note: `cargo check` cannot run in this sandbox because the egress proxy
blocks `cdn.pyke.io` (where `ort-sys` fetches its prebuilt ONNX Runtime
binary). The refactor is a pure code-move with no signature or call-site
semantic changes — verifiable by inspection. Please run
`cargo test --manifest-path src-tauri/Cargo.toml` locally before merge.

Files:
- src-tauri/src/semantic/runtime.rs:1-130: New file. Houses the five
  helper functions, the `CachedChunk` struct, and the static state
  (`EMBEDDER`, `VAULT_PATH`, `UNLOAD_GENERATION`,
  `EMBEDDER_IDLE_TIMEOUT_SECS`, `SEARCH_CACHE`), all `pub(crate)`.
- src-tauri/src/semantic/mod.rs:5: Added `pub mod runtime;`.
- src-tauri/src/commands/semantic.rs:1-21: Replaced local definitions
  with `use crate::semantic::runtime::{...}` import. Removed
  `std::sync::{Arc, Mutex}` and `std::sync::atomic::{AtomicU64, Ordering}`
  imports (no longer needed). `BUILD_LOCK` remains as the only static
  in this file.
- tasks/todo/personal-rag.md:1-29: New plan tracker for the RAG layer
  per CLAUDE.md plan-mode workflow; Task 1 marked complete.
