+++ src-tauri/src/semantic/mod.rs
@@ -2,4 +2,5 @@ pub mod chunker;
 pub mod embedder;
 pub mod filtering;
 pub mod model;
+pub mod runtime;
 pub mod types;
